+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.0.9
 - journal with docs
 - welcome screen
+- chat privacy
+- chat message style fix
 
 # 0.0.8
 First release